chore: move styles folder from components to src (#21)
