+++ packages/bootstrap/src/components/abs-button.js
@@ -1,6 +1,6 @@
 import { ButtonElement } from '@aybolit/core';
-import absButtonStyles from './styles/abs-button-css.js';
-import absThemeStyles from './styles/abs-theme-css.js';
+import absButtonStyles from '../styles/abs-button-css.js';
+import absThemeStyles from '../styles/abs-theme-css.js';
 
 export class AbsButton extends ButtonElement {
   static get styles() {
