+++ packages/bootstrap/src/components/abs-checkbox.js
@@ -1,6 +1,6 @@
 import { CheckboxElement } from '@aybolit/core';
-import absCheckboxStyles from './styles/abs-checkbox-css.js';
-import absThemeStyles from './styles/abs-theme-css.js';
+import absCheckboxStyles from '../styles/abs-checkbox-css.js';
+import absThemeStyles from '../styles/abs-theme-css.js';
 
 export class AbsCheckbox extends CheckboxElement {
   static get styles() {
