+++ packages/bootstrap/src/components/abs-progress.js
@@ -1,6 +1,6 @@
 import { ProgressElement } from '@aybolit/core';
-import absProgressStyles from './styles/abs-progress-css.js';
-import absThemeStyles from './styles/abs-theme-css.js';
+import absProgressStyles from '../styles/abs-progress-css.js';
+import absThemeStyles from '../styles/abs-theme-css.js';
 
 export class AbsProgress extends ProgressElement {
   static get styles() {
