+++ packages/bootstrap/src/components/abs-range.js
@@ -1,6 +1,6 @@
 import { RangeElement } from '@aybolit/core';
-import absRangeStyles from './styles/abs-range-css.js';
-import absThemeStyles from './styles/abs-theme-css.js';
+import absRangeStyles from '../styles/abs-range-css.js';
+import absThemeStyles from '../styles/abs-theme-css.js';
 
 export class AbsRange extends RangeElement {
   static get styles() {
