+++ packages/bootstrap/src/components/abs-switch.js
@@ -1,6 +1,6 @@
 import { SwitchElement } from '@aybolit/core';
-import absSwitchStyles from './styles/abs-switch-css.js';
-import absThemeStyles from './styles/abs-theme-css.js';
+import absSwitchStyles from '../styles/abs-switch-css.js';
+import absThemeStyles from '../styles/abs-theme-css.js';
 
 export class AbsSwitch extends SwitchElement {
   static get styles() {
