+++ packages/bulma/src/components/abu-button.js
@@ -1,6 +1,6 @@
 import { ButtonElement } from '@aybolit/core';
-import abuButtonStyles from './styles/abu-button-css.js';
-import abuThemeStyles from './styles/abu-theme-css.js';
+import abuButtonStyles from '../styles/abu-button-css.js';
+import abuThemeStyles from '../styles/abu-theme-css.js';
 
 export class AbuButton extends ButtonElement {
   static get styles() {
