+++ packages/bulma/src/components/abu-checkbox.js
@@ -1,6 +1,6 @@
 import { CheckboxElement } from '@aybolit/core';
-import abuCheckboxStyles from './styles/abu-checkbox-css.js';
-import abuThemeStyles from './styles/abu-theme-css.js';
+import abuCheckboxStyles from '../styles/abu-checkbox-css.js';
+import abuThemeStyles from '../styles/abu-theme-css.js';
 
 export class AbuCheckbox extends CheckboxElement {
   static get styles() {
