+++ packages/bulma/src/components/abu-progress.js
@@ -1,6 +1,6 @@
 import { ProgressElement } from '@aybolit/core';
-import abuProgressStyles from './styles/abu-progress-css.js';
-import abuThemeStyles from './styles/abu-theme-css.js';
+import abuProgressStyles from '../styles/abu-progress-css.js';
+import abuThemeStyles from '../styles/abu-theme-css.js';
 
 export class AbuProgress extends ProgressElement {
   static get styles() {
