+++ packages/bulma/src/components/abu-range.js
@@ -1,6 +1,6 @@
 import { RangeElement } from '@aybolit/core';
-import abuRangeStyles from './styles/abu-range-css.js';
-import abuThemeStyles from './styles/abu-theme-css.js';
+import abuRangeStyles from '../styles/abu-range-css.js';
+import abuThemeStyles from '../styles/abu-theme-css.js';
 
 export class AbuRange extends RangeElement {
   static get styles() {
