+++ packages/bulma/src/components/abu-switch.js
@@ -1,6 +1,6 @@
 import { SwitchElement } from '@aybolit/core';
-import abuSwitchStyles from './styles/abu-switch-css.js';
-import abuThemeStyles from './styles/abu-theme-css.js';
+import abuSwitchStyles from '../styles/abu-switch-css.js';
+import abuThemeStyles from '../styles/abu-theme-css.js';
 
 export class AbuSwitch extends SwitchElement {
   static get styles() {
