+++ packages/core/src/components/range.js
@@ -1,7 +1,7 @@
 import { LitElement, html } from 'lit-element';
 import { ifDefined } from 'lit-html/directives/if-defined';
 import { DelegateFocusMixin } from '../mixins/delegate-focus-mixin.js';
-import rangeBaseStyles from './styles/range-base-css.js';
+import rangeBaseStyles from '../styles/range-base-css.js';
 
 const isNumeric = n => !isNaN(parseFloat(n)) && isFinite(n);
