+++ packages/core/src/styles/.gitignore
