+++ packages/material/src/components/abm-button.js
@@ -1,5 +1,5 @@
 import { ButtonElement } from '@aybolit/core';
-import abmButtonStyles from './styles/abm-button-css.js';
+import abmButtonStyles from '../styles/abm-button-css.js';
 
 export class AbmButton extends ButtonElement {
   static get styles() {
