+++ packages/material/src/components/abm-checkbox.js
@@ -1,5 +1,5 @@
 import { CheckboxElement } from '@aybolit/core';
-import abmCheckboxStyles from './styles/abm-checkbox-css.js';
+import abmCheckboxStyles from '../styles/abm-checkbox-css.js';
 
 export class AbmCheckbox extends CheckboxElement {
   static get styles() {
