+++ packages/material/src/components/abm-progress.js
@@ -1,5 +1,5 @@
 import { ProgressElement } from '@aybolit/core';
-import abmProgressStyles from './styles/abm-progress-css.js';
+import abmProgressStyles from '../styles/abm-progress-css.js';
 
 export class AbmProgress extends ProgressElement {
   static get styles() {
