+++ packages/material/src/components/abm-range.js
@@ -1,5 +1,5 @@
 import { RangeElement } from '@aybolit/core';
-import abmRangeStyles from './styles/abm-range-css.js';
+import abmRangeStyles from '../styles/abm-range-css.js';
 
 export class AbmRange extends RangeElement {
   static get styles() {
