+++ packages/white-label/src/components/abw-button.js
@@ -1,5 +1,5 @@
 import { ButtonElement } from '@aybolit/core';
-import abwButtonStyles from './styles/abw-button-css.js';
+import abwButtonStyles from '../styles/abw-button-css.js';
 
 class AbwButton extends ButtonElement {
   static get styles() {
