+++ packages/white-label/src/components/abw-switch.js
@@ -1,5 +1,5 @@
 import { SwitchElement } from '@aybolit/core';
-import abwSwitchStyles from './styles/abw-switch-css.js';
+import abwSwitchStyles from '../styles/abw-switch-css.js';
 
 class AbwSwitch extends SwitchElement {
   static get styles() {
